requirements.txt
boto3
numpy
pandas
pymysql 
psycopg2
mysql.connector